Rename remix module to map
Remix is coming later
