+++ apps/core/templatetags/custom_filters.py
@@ -5,7 +5,7 @@ from django import template
 from django.utils.safestring import SafeString, mark_safe
 from django.utils.timesince import timesince, timeuntil
 
-from remix.markers import river_marker
+from map.markers import river_marker
 from river.models import River
 
 register = template.Library()
