+++ apps/map/views.py
@@ -1,12 +1,12 @@
 from django.views.generic import TemplateView
 
-from remix.markers import river_marker, case_study_marker
+from map.markers import river_marker, case_study_marker
 from resources.models import CaseStudy, HowTo
 from river.models import River
 
 
-class RemixMapView(TemplateView):
-    template_name = "remix/remix_map.html"
+class MapView(TemplateView):
+    template_name = "map/map.html"
 
     def get_context_data(self, **kwargs):
         context = super().get_context_data(**kwargs)
