+++ templates/partials/nav.html
@@ -92,8 +92,8 @@
                     </div>
                 </div>
 
-                <li class="p-4 pb-5 {% active_link 'text-purple' 'remix_map' %}">
-                    <a href="{% url 'remix_map' %}">Remix</a>
+                <li class="p-4 pb-5 {% active_link 'text-purple' 'map' %}">
+                    <a href="{% url 'map' %}">Map</a>
                 </li>
 
                 {% if user.is_authenticated %}
